API-IFとViewModel構築を二層のサービスに分割する途中
