+++ app/components/power/power-brand.component.ts
@@ -1,7 +1,7 @@
 import {Component,OnInit,Input} from 'angular2/core';
-import {Observable}     from 'rxjs/Observable';
-import {GearBrand}       from 'types/gear-brand';
-import {GearBrandService} from 'services/gear-brand.service';
+import {Gear,GearType,GearBrand,GearPower} from 'types/api';
+import {GearService,GearTypeService,GearBrandService,GearPowerService} from 'services/api.service';
+import {GearDataService} from 'services/data.service'
 
 @Component({
     selector: 'power-brand',
