+++ app/types/api.ts
@@ -1,0 +1,23 @@
+export class Gear {
+    id:number;
+    label:string;
+    main:number;
+    brand:number;
+    rarity:number;
+    type:number;
+}
+export class GearType {
+    id: number;
+    name: string;
+    label: string;
+}
+export class GearPower {
+    id: number;
+    label: string;
+}
+export class GearBrand {
+    id: number;
+    label: string;
+    strong: number;
+    weak: number;
+}
