--- app/types/gear-brand.ts
@@ -1,6 +1,0 @@
-export class GearBrand {
-    id: number;
-    label: string;
-    strong: number;
-    weak: number;
-}
