--- app/types/gear-power.ts
@@ -1,4 +1,0 @@
-export class GearPower {
-    id: number;
-    label: string;
-}
