--- app/types/gear-type.ts
@@ -1,5 +1,0 @@
-export class GearType {
-    id: number;
-    name: string;
-    label: string;
-}
