--- app/types/gear.ts
@@ -1,8 +1,0 @@
-export class Gear {
-    id:number;
-    label:string;
-    main:number;
-    brand:number;
-    rarity:number;
-    type:number;
-}
